# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Perfect Fit API",
        "description": "API documentation for Perfect Fit application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user details",
                "description": "Retrieve the current authenticated user's information",
                "operationId": "6c28a65da4d230795e7b7e3293dc1a61",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User information retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "Vietnam"
                                                },
                                                "province": {
                                                    "type": "string",
                                                    "example": "Ho Chi Minh"
                                                },
                                                "district": {
                                                    "type": "string",
                                                    "example": "District 1"
                                                },
                                                "ward": {
                                                    "type": "string",
                                                    "example": "Ward 1"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Main St"
                                                },
                                                "postal_code": {
                                                    "type": "string",
                                                    "example": "70000"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "images/upload/user/photo.jpg"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T10:00:00Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/update-info": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update current user information",
                "description": "Update the current authenticated user's profile information",
                "operationId": "f729b188194a53aa977fe4e97b003721",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Updated"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Vietnam"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Ho Chi Minh"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "District 1"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Ward 1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Updated St"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "70000"
                                    },
                                    "profile_photo_path": {
                                        "description": "Profile photo file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Updated"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhập thông tin thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token in the format: Bearer <token>"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Enter your API key in the format: Bearer <token>"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}